+++ user-operator/src/main/java/io/strimzi/operator/user/model/KafkaUserModel.java
@@ -45,6 +45,7 @@ public class KafkaUserModel {
     private static final Logger log = LogManager.getLogger(KafkaUserModel.class.getName());
 
     public static final String KEY_PASSWORD = "password";
+    public static final String KEY_SASL_JAAS_CONFIG = "sasl.jaas.config";
 
     protected final String namespace;
     protected final String name;
@@ -157,8 +158,9 @@ public class KafkaUserModel {
             data.put("user.password", userCertAndKey.storePasswordAsBase64String());
             return createSecret(data);
         } else if (authentication instanceof KafkaUserScramSha512ClientAuthentication) {
-            Map<String, String> data = new HashMap<>(1);
-            data.put(KafkaUserModel.KEY_PASSWORD, Base64.getEncoder().encodeToString(scramSha512Password.getBytes(StandardCharsets.US_ASCII)));
+            Map<String, String> data = new HashMap<>(2);
+            data.put(KafkaUserModel.KEY_PASSWORD, Base64.getEncoder().encodeToString(this.scramSha512Password.getBytes(StandardCharsets.US_ASCII)));
+            data.put(KafkaUserModel.KEY_SASL_JAAS_CONFIG, Base64.getEncoder().encodeToString(getSaslJsonConfig().getBytes(StandardCharsets.US_ASCII)));
             return createSecret(data);
         } else {
             return null;
@@ -395,6 +397,17 @@ public class KafkaUserModel {
     }
 
     /**
+     * Creates the JAAS configuration string for SASL SCRAM-SHA-512 authentication.
+     *
+     * @param username The SCRAM username.
+     * @param password The SCRAM password.
+     * @return The JAAS configuration string.
+     */
+    public static String getSaslJsonConfig(String username, String password) {
+        return "org.apache.kafka.common.security.scram.ScramLoginModule required username=\"" + username + "\" password=\"" + password + "\";";
+    }
+
+    /**
      * Gets the name of the User secret.
      *
      * @return The name of the user secret.
@@ -482,4 +495,14 @@ public class KafkaUserModel {
     public boolean isNoneUser()  {
         return authentication == null;
     }
+
+    /**
+     * Creates the JAAS configuration string for SASL SCRAM-SHA-512 authentication.
+     *
+     * @return The JAAS configuration string.
+     */
+    public String getSaslJsonConfig() {
+        return getSaslJsonConfig(getScramUserName(name), scramSha512Password);
+    }
+
 }